services/laboratory_service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pymongo==4.6.1
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
gunicorn==21.2.0
drf-yasg==1.21.7
django-cors-headers==4.3.1 